* Fix or toggle off warnings (mostly TF deprecation warnings)
* Update dependency to TF>=1.10.0
* Upgrade to version v0.2.2

+++ examples/transformer/transformer_main.py
@@ -22,8 +22,8 @@ import pickle
 import random
 import os
 import importlib
-from torchtext import data
 import tensorflow as tf
+from torchtext import data
 import texar as tx
 from texar.modules import TransformerEncoder, TransformerDecoder
 from texar.utils import transformer_utils
@@ -74,10 +74,10 @@ def main():
     batch_size = tf.shape(encoder_input)[0]
     # (text sequence length excluding padding)
     encoder_input_length = tf.reduce_sum(
-        1 - tf.to_int32(tf.equal(encoder_input, 0)), axis=1)
+        1 - tf.cast(tf.equal(encoder_input, 0), tf.int32), axis=1)
 
     labels = tf.placeholder(tf.int64, shape=(None, None))
-    is_target = tf.to_float(tf.not_equal(labels, 0))
+    is_target = tf.cast(tf.not_equal(labels, 0), tf.float32)
 
     global_step = tf.Variable(0, dtype=tf.int64, trainable=False)
     learning_rate = tf.placeholder(tf.float64, shape=(), name='lr')
